set up using gulp
- gulp does linting and building to /dist
- js now served out of dist
- made changes to listen to linter
- modules export
