+++ index.html
@@ -13,12 +13,11 @@
 
 	<!-- vendor assets -->
 	<script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/raphael/2.2.7/raphael.min.js"></script>
-	<!-- <script type="text/javascript" src="js/raphael/raphael.js"></script> -->
 
-	<script type="text/javascript" src="js/util.js"></script>
-	<script type="text/javascript" src="js/pythagoras-tree.js"></script>
-	<script type="text/javascript" src="js/sierpinski-triangle.js"></script>
-	<script type="text/javascript" src="js/sierpinski-arrowhead-curve.js"></script>
+	<script type="text/javascript" src="dist/util.js"></script>
+	<script type="text/javascript" src="dist/pythagoras-tree.js"></script>
+	<script type="text/javascript" src="dist/sierpinski-triangle.js"></script>
+	<script type="text/javascript" src="dist/sierpinski-arrowhead-curve.js"></script>
 
 
 	<link rel="stylesheet" href="css/fractals.css">
@@ -48,7 +47,7 @@
 		<p>The Sierpi&#324;ski Triangle can be generated by following the rule that every triangle has a smaller triangle drawn on each of its sides.</p>
 		<p>At each level, triangles are drawn with half the side length of the triangles drawn at the previous level.</p>
 		<br/>
-		<a href="https://github.com/aberke/fractals/blob/master/js/sierpinski-triangle.js" class="code-link">See the code that made this.</a>
+		<a href="https://github.com/aberke/fractals/blob/master/dist/sierpinski-triangle.js" class="code-link">See the code that made this.</a>
 		<br/>
 		<p>Another fractal that draws along the outline of the Sierpi&#324;ski Triangle is the Sierpi&#324;ski Arrowhead Curve</p>
 
@@ -78,7 +77,7 @@
 		<br/>
 		<p>The curve can be defined more formally with a <a class="bold" target="_blank" href="http://mathworld.wolfram.com/StringRewritingSystem.html">rewrite system</a>.</p>
 		<br/>
-		<a href="https://github.com/aberke/fractals/blob/master/js/sierpinski-arrowhead-curve.js" class="code-link">See the code that made this.</a>
+		<a href="https://github.com/aberke/fractals/blob/master/dist/sierpinski-arrowhead-curve.js" class="code-link">See the code that made this.</a>
 
 	</section>
 
@@ -98,7 +97,7 @@
 		Each level will then have 2 times as many squares as the previous level, and each level's squares will have side lengths that are &frac12;&radic;2 times the size of the previous level's.
 		<br/>
 		<br/>
-		<a href="https://github.com/aberke/fractals/blob/master/js/pythagoras-tree.js" class="code-link">See the code that made this.</a>
+		<a href="https://github.com/aberke/fractals/blob/master/dist/pythagoras-tree.js" class="code-link">See the code that made this.</a>
 		<br/>
 		Now consider a Pythagoras Tree where the lines that make up the squares of the tree are not lines, but replaced with other edges.  These edges may curve, but they still end up at the same corners the square's lines would have ended.
 		<br/>
@@ -146,12 +145,12 @@ var paper1 = new Raphael("canvas-1", "100%", TRIANGLE_CANVAS_HEIGHT);
 var paper2 = new Raphael("canvas-2", "100%", TRIANGLE_CANVAS_HEIGHT);
 
 var maxLevel = 6;
-var baseFractalFunction = getSierpinskiTriangle;
+var baseFractalFunction = sierpinskiTriangle.getSierpinskiTriangle;
 
 // draw the Sierpinski Triangle with a 'redraw button'
 var redrawSierpinskiTriangleFractals = function() {
 	paper1.clear();
-	var sierpinskiTriangleFractals = drawFractalRow(paper1, TRIANGLE_ROW_COUNT, maxLevel, getSierpinskiTriangle, baseFractalFunction);
+	var sierpinskiTriangleFractals = fractalsUtil.drawFractalRow(paper1, TRIANGLE_ROW_COUNT, maxLevel, sierpinskiTriangle.getSierpinskiTriangle, baseFractalFunction);
 }
 var redrawSierpinskiTriangleFractalsButton = document.getElementById("sierpinski-triangle-redraw-btn");
 redrawSierpinskiTriangleFractalsButton.onclick = redrawSierpinskiTriangleFractals;
@@ -161,7 +160,7 @@ redrawSierpinskiTriangleFractals();
 // draw the arrow head curves with a 'redraw' button
 var redrawSierpinskiArrowheadCurveFractals = function() {
 	paper2.clear();
-	var sierpinskiArrowheadCurveFractals = drawFractalRow(paper2, TRIANGLE_ROW_COUNT, 4, getSierpinskiArrowheadCurve, baseFractalFunction);
+	var sierpinskiArrowheadCurveFractals = fractalsUtil.drawFractalRow(paper2, TRIANGLE_ROW_COUNT, 4, sierpinskiArrowheadCurve.getSierpinskiArrowheadCurve, baseFractalFunction);
 }
 var redrawSierpinskiArrowheadCurveFractalsButton = document.getElementById("sierpinski-arrowhead-curve-redraw-btn");
 redrawSierpinskiArrowheadCurveFractalsButton.onclick = redrawSierpinskiArrowheadCurveFractals;
@@ -182,7 +181,7 @@ function handlePythagorasCanvas(containerElt, edgePathFunction) {
 	var sideLength = (5/3)*Math.min(paperWidth, paperHeight)/maxLevel;
 	var centerPoint = {X: paperWidth/2, Y: paperHeight - (1/2)*sideLength};
 
-	var pythagorasTreePathList = getPythagorasTree(centerPoint, sideLength, {
+	var pythagorasTreePathList = pythagorasTree.getPythagorasTree(centerPoint, sideLength, {
 		levels: maxLevel,
 		edgePathFunction: edgePathFunction,
 	});
@@ -199,7 +198,7 @@ function handlePythagorasCanvas(containerElt, edgePathFunction) {
 	function redrawPythagorasTree() {
 		paper.clear();
 		currentDrawLevel = 0;
-		drawPythagorasTree(paper, pythagorasTreePathList, 1000, drawPythagorasTreeCallback);
+		pythagorasTree.drawPythagorasTree(paper, pythagorasTreePathList, 1000, drawPythagorasTreeCallback);
 	}
 	redrawButtonElt.onclick = redrawPythagorasTree;
 	// draw the tree for the first time
@@ -230,11 +229,11 @@ function drawPythagorasTreeMultiple(paper, edgePathFunctions, drawCallback) {
 
 		var centerPoint = {X: centerPointX, Y: centerPointY};
 
-		var pythagorasTreePathList = getPythagorasTree(centerPoint, sideLength, {
+		var pythagorasTreePathList = pythagorasTree.getPythagorasTree(centerPoint, sideLength, {
 			levels: maxLevel,
 			edgePathFunction: edgePathFunctions[i],
 		});
-		drawPythagorasTree(paper, pythagorasTreePathList, 1000, drawCallback);
+		pythagorasTree.drawPythagorasTree(paper, pythagorasTreePathList, 1000, drawCallback);
 	}
 }
 
@@ -248,9 +247,9 @@ function redrawPythagorasTreeMultiple() {
 	paper4.clear();
 	var randomMultiplier = Math.random();
 	var randomEdgePathFunction = function(fromPoint, toPoint, centerPoint) {
-		return getCatmullRomPath(fromPoint, toPoint, centerPoint, randomMultiplier);
+		return pythagorasTree.getCatmullRomPath(fromPoint, toPoint, centerPoint, randomMultiplier);
 	};
-	var edgePathFunctions = [getEllipsePath, getCurvedPath, randomEdgePathFunction];
+	var edgePathFunctions = [pythagorasTree.getEllipsePath, pythagorasTree.getCurvedPath, randomEdgePathFunction];
 
 	// update level text element as the level increments	
 	var currentDrawLevel = 0;
