+++ package.json
@@ -1,0 +1,37 @@
+{
+  "name": "fractals",
+  "version": "0.1.0",
+  "description": "Fractals are patterns that are generated by repeating themselves..",
+  "dependencies": {},
+  "devDependencies": {
+    "babel-preset-env": "^1.4.0",
+    "gulp": "^3.9.1",
+    "gulp-babel": "^6.1.2",
+    "gulp-jshint": "^2.0.4",
+    "jshint": "^2.9.4"
+  },
+  "jshintConfig": {
+    "undef": true,
+    "unused": true,
+    "esversion": 6
+  },
+  "scripts": {},
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/aberke/fractals.git"
+  },
+  "keywords": [
+    "fractals",
+    "trees",
+    "shapes",
+    "L-system",
+    "animation",
+    "raphaelJS"
+  ],
+  "author": "Alex Berke",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/aberke/fractals/issues"
+  },
+  "homepage": "https://aberke.github.io/fractals"
+}
